requirements.txt
Appium-Python-Client==3.1.1
selenium==4.25.0
pillow==10.3.0
psutil==5.9.8
requests==2.32.3
colorama==0.4.6
PySimpleGUI==5.0.8.3
